chai_research/qlora/requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@21d127334ebbcdd4f2f0695a5b36753dae6d747b
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.41.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
datasets==2.14.4
dill==0.3.7
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
peft @ git+https://github.com/huggingface/peft.git@312d294fddba3ef51d1f7df1e7fa1793ca0a5d05
psutil==5.9.5
pyarrow==12.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers.git@4a27c13f1eee26393d60d381e500e1a61970e8ee
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xxhash==3.3.0
yarl==1.9.2
